+++ work_spacer/lib/models/room_reservation.dart
@@ -1,17 +1,16 @@
 import 'package:flutter/foundation.dart';
+import 'package:work_spacer/models/reservation.dart';
 import 'package:work_spacer/models/room.dart';
 
 @immutable
-class RoomReservation {
+class RoomReservation extends Reservation {
   final Room room;
-  final DateTime startDate;
-  final int duration;
-  final int idEmployee;
+
 
   const RoomReservation({
     required this.room,
-    required this.startDate,
-    required this.duration,
-    required this.idEmployee,
+    required super.startDate,
+    required super.duration,
+    required super.idEmployee,
   });
 }