+++ work_spacer/lib/screens/home/employee_home_content.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 import 'package:work_spacer/screens/employee/desks/desk_search_screen.dart';
 import 'package:work_spacer/screens/employee/reservations/employee_reservations_screen.dart';
 import 'package:work_spacer/screens/employee/rooms/room_search_screen.dart';
+import 'components/home_screen_button.dart';
 
 class EmployeeHomeContent extends StatelessWidget {
   const EmployeeHomeContent({
@@ -10,62 +12,35 @@ class EmployeeHomeContent extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Card(
-          child: ListTile(
-            leading: const Icon(
-              Icons.event_available_rounded,
-              color: Colors.blue,
-            ),
-            title: const Text(
-              'My reservations',
-              style: TextStyle(
-                color: Colors.blue,
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-            onTap: () {
-              Navigator.restorablePushNamed(
-                context,
-                EmployeeReservationsScreen.routeName,
-              );
-            },
-          ),
+    return Container(
+      alignment: Alignment.center,
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+        HomeScreenButton(
+          backgroundColor: Colors.deepOrange.shade200,
+          contentColor: Colors.white,
+          icon: Icons.event_available_rounded,
+          title: "My reservations",
+          navRouteName: EmployeeReservationsScreen.routeName,
         ),
-        const SizedBox(height: 8),
-        Card(
-          child: ListTile(
-            leading: const Icon(
-              Icons.desktop_windows_rounded,
-              color: Colors.blue,
-            ),
-            title: const Text('Desks'),
-            onTap: () {
-              Navigator.restorablePushNamed(
-                context,
-                DeskSearchScreen.routeName,
-              );
-            },
-          ),
+        const SizedBox(height: 32),
+        HomeScreenButton(
+          backgroundColor: Colors.teal.shade300,
+          contentColor: Colors.teal.shade900,
+          icon: Icons.desktop_windows_outlined,
+          title: "Desks",
+          navRouteName: DeskSearchScreen.routeName,
         ),
-        const SizedBox(height: 8),
-        Card(
-          child: ListTile(
-            leading: const Icon(
-              Icons.meeting_room_rounded,
-              color: Colors.blue,
-            ),
-            title: const Text('Rooms'),
-            onTap: () {
-              Navigator.restorablePushNamed(
-                context,
-                RoomSearchScreen.routeName,
-              );
-            },
-          ),
+        const SizedBox(height: 32),
+        HomeScreenButton(
+          backgroundColor: Colors.teal.shade300,
+          contentColor: Colors.teal.shade900,
+          icon: Icons.meeting_room_outlined,
+          title: "Conference rooms",
+          navRouteName: RoomSearchScreen.routeName,
         ),
-      ],
+      ]),
     );
   }
 }